requirements.txt
#python==3.6
pyshp==1.2.12
numpy==1.15.3
matplotlib==3.0.1
pandas==0.23.4
PyYAML==3.13
#seaborn==0.9.0